+++ CHANGELOG.rst
@@ -25,6 +25,8 @@ Fixed
 ^^^^^
 - Stubs resolver in some cases failing with maximum recursion error (`#770
   <https://github.com/omni-us/jsonargparse/pull/770>`__).
+- ``dataclass`` with default failing when ``validate_defaults=True`` (`#771
+  <https://github.com/omni-us/jsonargparse/pull/771>`__).
 
 
 v4.41.0 (2025-09-04)